+++ demo/src/test/java/Launcher.java
@@ -3,7 +3,6 @@ import javafx.application.Application;
 public class Launcher {
 
 	public static void main(String[] args) {
-		System.setProperty("prism.text", "t2k");
-		Application.launch(Playground.class, args);
+		Application.launch(MagnifierTest.class, args);
 	}
 }