+++ demo/src/test/java/selection/SingleSelectionModelTests.java
@@ -2,7 +2,7 @@ package selection;
 
 import io.github.palexdev.materialfx.bindings.BiBindingManager;
 import io.github.palexdev.materialfx.bindings.BindingManager;
-import io.github.palexdev.materialfx.demo.model.SimplePerson;
+import io.github.palexdev.materialfx.demo.model.Person;
 import io.github.palexdev.materialfx.selection.SingleSelectionModel;
 import javafx.beans.property.*;
 import javafx.collections.FXCollections;
@@ -18,33 +18,33 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(ApplicationExtension.class)
 public class SingleSelectionModelTests {
-	private final ListProperty<SimplePerson> people1 = new SimpleListProperty<>(
+	private final ListProperty<Person> people1 = new SimpleListProperty<>(
 			FXCollections.observableArrayList(
-					new SimplePerson("Jack"),
-					new SimplePerson("Mark"),
-					new SimplePerson("Linda"),
-					new SimplePerson("Marty"),
-					new SimplePerson("Lily"),
-					new SimplePerson("Sam"))
+					new Person("Jack"),
+					new Person("Mark"),
+					new Person("Linda"),
+					new Person("Marty"),
+					new Person("Lily"),
+					new Person("Sam"))
 	);
 
-	private final ListProperty<SimplePerson> people2 = new SimpleListProperty<>(
+	private final ListProperty<Person> people2 = new SimpleListProperty<>(
 			FXCollections.observableArrayList(
-					new SimplePerson("Mark"),
-					new SimplePerson("Roberto"),
-					new SimplePerson("Alex"),
-					new SimplePerson("Samantha"),
-					new SimplePerson("Elyse"),
-					new SimplePerson("Mark"),
-					new SimplePerson("Sam"),
-					new SimplePerson("Jennifer"),
-					new SimplePerson("Alex"),
-					new SimplePerson("Rocky"),
-					new SimplePerson("Phil"))
+					new Person("Mark"),
+					new Person("Roberto"),
+					new Person("Alex"),
+					new Person("Samantha"),
+					new Person("Elyse"),
+					new Person("Mark"),
+					new Person("Sam"),
+					new Person("Jennifer"),
+					new Person("Alex"),
+					new Person("Rocky"),
+					new Person("Phil"))
 	);
 
-	private final SingleSelectionModel<SimplePerson> model1 = new SingleSelectionModel<>(people1);
-	private final SingleSelectionModel<SimplePerson> model2 = new SingleSelectionModel<>(people2);
+	private final SingleSelectionModel<Person> model1 = new SingleSelectionModel<>(people1);
+	private final SingleSelectionModel<Person> model2 = new SingleSelectionModel<>(people2);
 
 	@AfterEach
 	public void checkManagers() {
@@ -79,14 +79,14 @@ public class SingleSelectionModelTests {
 
 	@Test
 	public void testItemSelection1() {
-		model1.selectItem(new SimplePerson("Mark"));
+		model1.selectItem(new Person("Mark"));
 		assertEquals(1, model1.getSelectedIndex());
 		assertEquals("Mark", model1.getSelectedItem().getName());
 	}
 
 	@Test
 	public void testItemSelection2() {
-		assertThrows(IllegalArgumentException.class, () -> model1.selectItem(new SimplePerson("Unexisting")));
+		assertThrows(IllegalArgumentException.class, () -> model1.selectItem(new Person("Unexisting")));
 	}
 
 	@Test
@@ -196,7 +196,7 @@ public class SingleSelectionModelTests {
 		assertEquals("Sam", model1.getSelectedItem().getName());
 		assertEquals(5, property.get());
 
-		model1.selectItem(new SimplePerson("Lily"));
+		model1.selectItem(new Person("Lily"));
 		assertEquals(4, model1.getSelectedIndex());
 		assertEquals(4, property.get());
 	}
@@ -276,20 +276,20 @@ public class SingleSelectionModelTests {
 
 	@Test
 	public void testBindItemProperty1() {
-		ObjectProperty<SimplePerson> property = new SimpleObjectProperty<>();
-		model1.bindItem(property, simplePerson -> model1.getUnmodifiableItems().indexOf(simplePerson));
+		ObjectProperty<Person> property = new SimpleObjectProperty<>();
+		model1.bindItem(property, Person -> model1.getUnmodifiableItems().indexOf(Person));
 
-		property.set(new SimplePerson("Mark"));
+		property.set(new Person("Mark"));
 		assertEquals(1, model1.getSelectedIndex());
 		assertEquals("Mark", model1.getSelectedItem().getName());
 	}
 
 	@Test
 	public void testBindItemProperty2() {
-		ObjectProperty<SimplePerson> property = new SimpleObjectProperty<>();
-		model2.bindItem(property, simplePerson -> model2.getUnmodifiableItems().indexOf(simplePerson));
+		ObjectProperty<Person> property = new SimpleObjectProperty<>();
+		model2.bindItem(property, Person -> model2.getUnmodifiableItems().indexOf(Person));
 
-		property.set(new SimplePerson("Alex"));
+		property.set(new Person("Alex"));
 		assertEquals(2, model2.getSelectedIndex());
 		assertEquals("Alex", model2.getSelectedItem().getName());
 	}
@@ -299,10 +299,10 @@ public class SingleSelectionModelTests {
 		AtomicReference<Throwable> reference = new AtomicReference<>();
 		Thread.setDefaultUncaughtExceptionHandler((t, e) -> reference.set(e));
 
-		ObjectProperty<SimplePerson> property = new SimpleObjectProperty<>();
-		model2.bindItem(property, simplePerson -> model2.getUnmodifiableItems().indexOf(simplePerson));
+		ObjectProperty<Person> property = new SimpleObjectProperty<>();
+		model2.bindItem(property, Person -> model2.getUnmodifiableItems().indexOf(Person));
 
-		property.set(new SimplePerson("Unexisting"));
+		property.set(new Person("Unexisting"));
 		assertEquals("Unexisting", property.get().getName());
 		assertEquals(-1, model2.getSelectedIndex());
 		assertNull(model2.getSelectedItem());
@@ -312,7 +312,7 @@ public class SingleSelectionModelTests {
 	@Test
 	public void testBindItem1() {
 		model1.bindItem(model2);
-		model2.selectItem(new SimplePerson("Mark"));
+		model2.selectItem(new Person("Mark"));
 		assertEquals(1, model1.getSelectedIndex());
 		assertEquals(0, model2.getSelectedIndex());
 	}
@@ -320,7 +320,7 @@ public class SingleSelectionModelTests {
 	@Test
 	public void testBindItem2() {
 		model1.bindItem(model2);
-		assertThrows(IllegalStateException.class, () -> model1.selectItem(new SimplePerson("Mark")));
+		assertThrows(IllegalStateException.class, () -> model1.selectItem(new Person("Mark")));
 	}
 
 	@Test
@@ -329,7 +329,7 @@ public class SingleSelectionModelTests {
 		Thread.setDefaultUncaughtExceptionHandler((t, e) -> reference.set(e));
 
 		model1.bindItem(model2);
-		model2.selectItem(new SimplePerson("Alex"));
+		model2.selectItem(new Person("Alex"));
 		assertEquals(-1, model1.getSelectedIndex());
 		assertNull(model1.getSelectedItem());
 		assertEquals(2, model2.getSelectedIndex());
@@ -339,10 +339,10 @@ public class SingleSelectionModelTests {
 
 	@Test
 	public void testBindItemPropertyBidirectional1() {
-		ObjectProperty<SimplePerson> property = new SimpleObjectProperty<>();
-		model1.bindItemBidirectional(property, simplePerson -> model1.getUnmodifiableItems().indexOf(simplePerson), (clearing, simplePerson, other) -> other.setValue(simplePerson));
+		ObjectProperty<Person> property = new SimpleObjectProperty<>();
+		model1.bindItemBidirectional(property, Person -> model1.getUnmodifiableItems().indexOf(Person), (clearing, Person, other) -> other.setValue(Person));
 
-		property.set(new SimplePerson("Linda"));
+		property.set(new Person("Linda"));
 		assertEquals(2, model1.getSelectedIndex());
 		assertEquals("Linda", model1.getSelectedItem().getName());
 	}
@@ -352,22 +352,22 @@ public class SingleSelectionModelTests {
 		AtomicReference<Throwable> reference = new AtomicReference<>();
 		Thread.setDefaultUncaughtExceptionHandler((t, e) -> reference.set(e));
 
-		ObjectProperty<SimplePerson> property = new SimpleObjectProperty<>();
-		model1.bindItemBidirectional(property, simplePerson -> model1.getUnmodifiableItems().indexOf(simplePerson), (clearing, simplePerson, other) -> other.setValue(simplePerson));
+		ObjectProperty<Person> property = new SimpleObjectProperty<>();
+		model1.bindItemBidirectional(property, Person -> model1.getUnmodifiableItems().indexOf(Person), (clearing, Person, other) -> other.setValue(Person));
 
-		property.set(new SimplePerson("Unexisting"));
+		property.set(new Person("Unexisting"));
 		assertNotNull(reference.get());
 	}
 
 	@Test
 	public void testBindItemPropertyBidirectional3() {
-		ObjectProperty<SimplePerson> property = new SimpleObjectProperty<>();
-		model1.bindItemBidirectional(property, simplePerson -> model1.getUnmodifiableItems().indexOf(simplePerson), (clearing, simplePerson, other) -> other.setValue(simplePerson));
+		ObjectProperty<Person> property = new SimpleObjectProperty<>();
+		model1.bindItemBidirectional(property, Person -> model1.getUnmodifiableItems().indexOf(Person), (clearing, Person, other) -> other.setValue(Person));
 
-		property.set(new SimplePerson("Mark"));
+		property.set(new Person("Mark"));
 		assertEquals(1, model1.getSelectedIndex());
 
-		model1.selectItem(new SimplePerson("Linda"));
+		model1.selectItem(new Person("Linda"));
 		assertEquals(2, model1.getSelectedIndex());
 		assertEquals("Linda", model1.getSelectedItem().getName());
 		assertEquals("Linda", property.get().getName());
@@ -377,11 +377,11 @@ public class SingleSelectionModelTests {
 	public void testBindItemBidirectional1() {
 		model1.bindItemBidirectional(model2);
 
-		model2.selectItem(new SimplePerson("Mark"));
+		model2.selectItem(new Person("Mark"));
 		assertEquals(1, model1.getSelectedIndex());
 		assertEquals(0, model2.getSelectedIndex());
 
-		model1.selectItem(new SimplePerson("Sam"));
+		model1.selectItem(new Person("Sam"));
 		assertEquals(5, model1.getSelectedIndex());
 		assertEquals(6, model2.getSelectedIndex());
 	}
@@ -392,11 +392,11 @@ public class SingleSelectionModelTests {
 		Thread.setDefaultUncaughtExceptionHandler((t, e) -> reference.set(e));
 		model1.bindItemBidirectional(model2);
 
-		model1.selectItem(new SimplePerson("Sam"));
+		model1.selectItem(new Person("Sam"));
 		assertEquals(5, model1.getSelectedIndex());
 		assertEquals(6, model2.getSelectedIndex());
 
-		model2.selectItem(new SimplePerson("Alex"));
+		model2.selectItem(new Person("Alex"));
 		assertNotNull(reference.get());
 		assertEquals(5, model1.getSelectedIndex());
 		assertEquals("Sam", model1.getSelectedItem().getName());
@@ -410,14 +410,14 @@ public class SingleSelectionModelTests {
 		Thread.setDefaultUncaughtExceptionHandler((t, e) -> reference.set(e));
 		model1.bindItemBidirectional(model2);
 
-		model2.selectItem(new SimplePerson("Alex"));
+		model2.selectItem(new Person("Alex"));
 		assertNotNull(reference.get());
 		assertEquals(-1, model1.getSelectedIndex());
 		assertNull(model1.getSelectedItem());
 		assertEquals(2, model2.getSelectedIndex());
 		assertEquals("Alex", model2.getSelectedItem().getName());
 
-		model1.selectItem(new SimplePerson("Sam"));
+		model1.selectItem(new Person("Sam"));
 		assertEquals(5, model1.getSelectedIndex());
 		assertEquals(6, model2.getSelectedIndex());
 	}
@@ -452,7 +452,7 @@ public class SingleSelectionModelTests {
 	@Test
 	public void testClearSelection4() {
 		model1.bindItemBidirectional(model2);
-		model2.selectItem(new SimplePerson("Mark"));
+		model2.selectItem(new Person("Mark"));
 
 		assertEquals(1, model1.getSelectedIndex());
 		assertEquals("Mark", model1.getSelectedItem().getName());
@@ -490,7 +490,7 @@ public class SingleSelectionModelTests {
 
 	@Test
 	public void testClearSelection6() {
-		ObjectProperty<SimplePerson> property = new SimpleObjectProperty<>();
+		ObjectProperty<Person> property = new SimpleObjectProperty<>();
 		model1.bindItemBidirectional(property, v -> model1.getUnmodifiableItems().indexOf(v), (clearing, v, otherProperty) -> property.set(v));
 		model1.selectIndex(1);