+++ materialfx/src/main/java/io/github/palexdev/materialfx/controls/MFXProgressBar.java
@@ -157,7 +157,7 @@ public class MFXProgressBar extends ProgressBar {
 	}
 
 	//================================================================================
-	// CssMetaData
+	// CSSMetaData
 	//================================================================================
 	private static class StyleableProperties {
 		private static final List<CssMetaData<? extends Styleable, ?>> cssMetaDataList;