+++ materialfx/src/main/java/io/github/palexdev/materialfx/controls/MFXToggleButton.java
@@ -315,7 +315,7 @@ public class MFXToggleButton extends Labeled implements Toggle, MFXLabeled {
 	}
 
 	//================================================================================
-	// CssMetaData
+	// CSSMetaData
 	//================================================================================
 	private static class StyleableProperties {
 		private static final List<CssMetaData<? extends Styleable, ?>> cssMetaDataList;