+++ materialfx/src/main/java/io/github/palexdev/materialfx/utils/others/FunctionalStringConverter.java
@@ -68,4 +68,42 @@ public interface FunctionalStringConverter<T> {
 			}
 		};
 	}
+
+	/**
+	 * @return a new {@link StringConverter} which is only capable of converting a String
+	 * to an object of type T
+	 * @throws UnsupportedOperationException when using the toString(T) method
+	 */
+	static <T> StringConverter<T> from(Function<String, T> fsFunction) {
+		return new StringConverter<>() {
+			@Override
+			public String toString(T t) {
+				throw new UnsupportedOperationException();
+			}
+
+			@Override
+			public T fromString(String string) {
+				return fsFunction.apply(string);
+			}
+		};
+	}
+
+	/**
+	 * @return a new {@link StringConverter} which is only capable of converting an object
+	 * of type T to a String
+	 * @throws UnsupportedOperationException when using the fromString(String) method
+	 */
+	static <T> StringConverter<T> to(Function<T, String> tsFunction) {
+		return new StringConverter<>() {
+			@Override
+			public String toString(T t) {
+				return tsFunction.apply(t);
+			}
+
+			@Override
+			public T fromString(String string) {
+				throw new UnsupportedOperationException();
+			}
+		};
+	}
 }